feat: services grid
Moved ServiceIcon to separate component
CSS props ordering
